+++ sdk/storage/Azure.Storage.Queues/CHANGELOG.md
@@ -1,7 +1,7 @@
 # Release History
 
 ## 12.5.0-preview.2 (Unreleased)
-
+- Fixed a bug where QueueServiceClient.SetProperties and QueueService.GetProperties where the creating/parsing XML Service Queue Properties CorsRules incorrectly causing Invalid XML Errors
 
 ## 12.5.0-preview.1 (2020-09-30)
 - This preview contains bug fixes to improve quality.